requirements.txt
mcp
kiteconnect
python-dotenv
pydantic